+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Rabbit Sky Client Changelog
 
+## Version 0.1.5.1
+- This version is compatible with Server v0.1.5
+- Adding gamepad controller (Beta)
+
 ## Version 0.1.5
 - Skip versioning so it's the same as server.
 - Improve bandwidth usage, using formatted message instead of CSV.